requirements.txt
pytubefix==9.2.0
python-telegram-bot==20.3
youtube-po-token-generator==0.1.3
requests
